move feature flags out of engine
