+++ packages/engine/src/EngineModule.ts
@@ -26,7 +26,6 @@ Infinite Reality Engine. All Rights Reserved.
 /** World Module */
 import '@ir-engine/spatial'
 
-export * from './FeatureFlagsState'
 export * from './avatar/AvatarModule'
 export * from './interaction/InteractionModule'
 export * from './interaction/MediaModule'
